Fixes for libjpeg compile (nw)

+++ 3rdparty/libjpeg/jmorecfg.h
@@ -162,7 +162,7 @@ typedef short INT16;
 #ifndef _BASETSD_H_		/* Microsoft defines it in basetsd.h */
 #ifndef _BASETSD_H		/* MinGW is slightly different */
 #ifndef QGLOBAL_H		/* Qt defines it in qglobal.h */
-typedef long INT32;
+typedef int INT32;
 #endif
 #endif
 #endif
